// packages/antdv-next/src/modal/interface.ts
export type ModalClassNamesType = SemanticClassNamesType<ModalProps, SemanticName>

export type ModalStylesType = SemanticStylesType<ModalProps, SemanticName>

interface ModalCommonProps
  extends Omit<
    DialogProps,
    'footer'
    | 'width'
    | 'onClose'
    | 'animation'
    | 'maskAnimation'
    | 'transitionName'
    | 'maskTransitionName'
    | 'mask'
    | 'classNames'
    | 'styles'
    | 'modalRender'
  >
{
  footer?: VueNode | ((params: { originNode: VueNode, extra: { OkBtn: any, CancelBtn: any } }) => any)
  closable?: boolean | (Exclude<ClosableType, boolean> & { onClose?: () => void, afterClose?: () => void })
  classes?: ModalClassNamesType
  styles?: ModalStylesType
}

type getContainerFunc = () => HTMLElement

export interface ModalProps extends ModalCommonProps {
  /** Whether the modal dialog is visible or not */
  open?: boolean
  /** Whether to apply loading visual effect for OK button or not */
  confirmLoading?: boolean
  /** The modal dialog's title */
  title?: VueNode
  afterClose?: () => void
  /** Callback when the animation ends when Modal is turned on and off */
  afterOpenChange?: (open: boolean) => void
  /** Centered Modal */
  centered?: boolean
  /** Width of the modal dialog */
  width?: string | number | Partial<Record<Breakpoint, string | number>>
  /** Text of the OK button */
  okText?: VueNode
  /** Button `type` of the OK button */
  okType?: LegacyButtonType
  /** Text of the Cancel button */
  cancelText?: VueNode
  /** Whether to close the modal dialog when the mask (area outside the modal) is clicked */
  maskClosable?: boolean
  /** Force render Modal */
  forceRender?: boolean
  okButtonProps?: ButtonProps
  cancelButtonProps?: ButtonProps
  /** @deprecated Please use `destroyOnHidden` instead */
  destroyOnClose?: boolean
  /**
   * @since 5.25.0
   */
  destroyOnHidden?: boolean
  wrapClassName?: string
  maskTransitionName?: string
  transitionName?: string
  rootClass?: string
  rootStyle?: CSSProperties
  getContainer?: string | HTMLElement | getContainerFunc | false
  zIndex?: number
  /** @deprecated Please use `styles.body` instead */
  bodyStyle?: CSSProperties
  /** @deprecated Please use `styles.mask` instead */
  maskStyle?: CSSProperties
  mask?: MaskType
  keyboard?: boolean
  wrapProps?: any
  prefixCls?: string
  closeIcon?: VueNode
  modalRender?: (node: any) => any
  focusTriggerAfterClose?: boolean
  mousePosition?: MousePosition
  /**
   * @since 5.18.0
   */
  loading?: boolean
}

export interface ModalEmits {
  /** Specify a function that will be called when a user clicks the OK button */
  ok: (e: MouseEvent) => void
  /** Specify a function that will be called when a user clicks mask, close button on top right or Cancel button */
  cancel: (e: MouseEvent) => void
  close: () => void
  [key: string]: (...args: any[]) => void
}

export interface ModalSlots {
  title?: () => any
  okText?: () => any
  cancelText?: () => any
  closeIcon?: () => any
  modalRender?: (node: any) => any
  footer?: (params: { originNode: VueNode, extra: { OkBtn: any, CancelBtn: any } }) => any
  default?: () => any
}
// ...
export type MousePosition = { x: number, y: number } | null